pyproject.toml
[tool.poetry]
name = "crud"
version = "0.1.0"
description = ""
authors = ["donkey4hot <tractorist666@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
pillow = "^9.5.0"
drf-spectacular = "^0.26.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
